pyproject.toml
[project]
name = "alice"
version = "0.3.0"
description = "Alice — the machine that runs her: agentic kernel, speaking daemon, thinking wake, observability viewer."
authors = [
    { name = "jcronq", email = "jcronq@gmail.com" }
]
# Viewer needs 3.11 (uses asyncio.timeout in narrative.py + fastapi features).
# Speaking's packages target 3.10+ but python 3.13 in the container covers both.
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]

dependencies = [
    # Agent kernel + hemispheres.
    "claude-agent-sdk>=0.1.66",
    "httpx>=0.28.1",
    # Per-transport markdown rendering (alice_speaking.render).
    "mistune>=3.0",
    # Address book / principals.yaml (alice_speaking.principals).
    "pyyaml>=6.0",
    # Discord transport (alice_speaking.transports.discord).
    "discord.py>=2.4",
    # Viewer.
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "sse-starlette>=2.0",
    # A2A protocol server (alice_speaking.transports.a2a). The
    # http-server extra pulls uvicorn + starlette; we reference 1.0+
    # APIs (new_task_from_user_message, AgentInterface,
    # protocol_version="1.0") and ship a workaround for issue #1011
    # specific to 1.0.x — must match the worker Dockerfile pin.
    "a2a-sdk[http-server]>=1.0.2",
]

[project.scripts]
alice-viewer = "alice_viewer.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Nine sibling Python packages under src/:
#   alice_core      — agentic kernel (SDK lifecycle + observability + auth)
#   alice_indexer   — vault SQLite/FTS5 indexer (CLI; cortex-memory tooling)
#   alice_pi        — Kernel impl backed by pi-coding-agent subprocess
#   alice_prompts   — Jinja2 templates the runtime renders into model prompts
#   alice_skills    — registry over .claude/skills/ markdown (plan 07)
#   alice_speaking  — Signal daemon (composes kernel with producers/handlers/outbox)
#   alice_thinking  — one-shot wake (python -m alice_thinking)
#   alice_viewer    — FastAPI observability UI (serves /state logs)
#   alice_watchers  — one-shot pollers feeding inner/notes/ (github, …)
[tool.hatch.build.targets.wheel]
packages = [
    "src/alice_core",
    "src/alice_indexer",
    "src/alice_pi",
    "src/alice_prompts",
    "src/alice_skills",
    "src/alice_speaking",
    "src/alice_thinking",
    "src/alice_viewer",
    "src/alice_watchers",
]

# alice_prompts ships .md.j2 template files alongside its Python
# modules; hatch needs the include-pattern to bundle them in the
# wheel. alice_skills bundles its runtime-default SKILL.md files
# the same way (plan 07 phase 6).
[tool.hatch.build.targets.wheel.force-include]
"src/alice_prompts/templates" = "alice_prompts/templates"
"src/alice_skills/defaults" = "alice_skills/defaults"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "strict"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.mypy]
# Lenient: don't fail on missing third-party stubs. Tighten later.
ignore_missing_imports = true
